Update separation of CRUD functionalities
