+++ client/src/App.tsx
@@ -5,12 +5,8 @@ import './App.css';
 import { Route, Routes } from 'react-router-dom';
 // We import all the components we need in our app
 import Navbarr from "./components/navbar";
-//import RecordList from "./components/recordList";
-import Edit from "./components/projects_all/edit";
-// import Create from "./components/mern/create";
 import Home from "./components/home";
 import AboutMe from "./components/about";
-import Mern from "./components/projects_all/mern";
 import Contact from "./components/contact";
 import Skills from "./components/skills";
 import Projects from "./components/projects";
@@ -19,7 +15,7 @@ import AwardsProj from './components/projects_all/awards';
 import SmsProj from './components/projects_all/sms';
 import SmsProj2 from './components/projects_all/sms2';
 import DilsProj from './components/projects_all/dils';
-import MernProj from './components/projects_all/mernx';
+import MernProj from './components/projects_all/mern';
 //import useScript from "./components/hooks";
 //import Footer from './components/footer';
 
@@ -27,13 +23,9 @@ function App() {
   return (
     <>
       <Navbarr />
-      <Routes>
-        {/* <Route exact path="/" element={<RecordList />} />*/}
-        <Route path="/" element={<Home/>}/>
-        <Route path="/edit/:id" element={<Edit/>} />
-        {/* <Route path="/create" element={<Create/>} />         */}
-        <Route path="/about" element={<AboutMe/>}/>
-        <Route path="/projects_all/mern" element={<Mern/>}/>
+      <Routes>     
+        <Route path="/" element={<Home/>}/>          
+        <Route path="/about" element={<AboutMe/>}/>       
         <Route path="/contact" element={<Contact/>}/>
         <Route path="/skills" element={<Skills/>}/>
         <Route path="/projects" element={<Projects/>}/> 
@@ -42,7 +34,7 @@ function App() {
         <Route path="/projects_all/sms" element={<SmsProj/>}/> 
         <Route path="/projects_all/sms2" element={<SmsProj2/>}/> 
         <Route path="/projects_all/dils" element={<DilsProj/>}/>     
-        <Route path="/projects_all/mernx" element={<MernProj/>}/>                
+        <Route path="/projects_all/mern" element={<MernProj/>}/>                
       </Routes> 
       {/* <Footer/>   */}
     </>
